refactor(dashboard): extract widget sizing policy module and split oversized board tests
