+++ ui/src/components/board/board-widget-cell-mcp-app.test.ts
@@ -41,6 +41,8 @@ function callbacks(overrides: Partial<BoardWidgetCellCallbacks> = {}): BoardWidg
     resizePointerDown: vi.fn(),
     moveToTab: noAction,
     resizeTo: noAction,
+    setHeightMode: noAction,
+    reportContentHeight: vi.fn(),
     remove: noAction,
     nudge: noAction,
     focus: vi.fn(),
